backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.10.6
pydantic-settings==2.7.1
httpx==0.27.2
supabase==2.10.0
python-dateutil==2.9.0.post0
